split out Sentence and Wordblock
